get rid of protobuf_conversions/log_types.rs
