+++ crates/store/re_protos/src/protobuf_conversions/mod.rs
@@ -1,2 +1,1 @@
 mod arrow;
-mod log_types;
